+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomainImpl.kt
@@ -61,20 +61,22 @@ private const val CHANNEL_LIMIT = 30
 internal val gson = Gson()
 
 /**
- * The Chat Repository exposes livedata objects to make it easier to build your chat UI.
+ * The Chat Domain exposes livedata objects to make it easier to build your chat UI.
  * It intercepts the various low level events to ensure data stays in sync.
  * Offline storage is handled using Room
  *
  * A different Room database is used for different users. That's why it's mandatory to specify the user id when
  * initializing the ChatRepository
  *
- * repo.channel(type, id) returns a repo object with channel specific livedata object
- * repo.queryChannels(query) returns a livedata object for the specific queryChannels query
+ * chatDomain.channel(type, id) returns a repo object with channel specific livedata object
+ * chatDomain.queryChannels(query) returns a livedata object for the specific queryChannels query
  *
- * repo.online livedata object indicates if you're online or not
- * repo.totalUnreadCount livedata object returns the current unread count for this user
- * repo.channelUnreadCount livedata object returns the number of unread channels for this user
- * repo.errorEvents events for errors that happen while interacting with the chat
+ * chatDomain.online livedata object indicates if you're online or not
+ * chatDomain.totalUnreadCount livedata object returns the current unread count for this user
+ * chatDomain.muted the list of muted users
+ * chatDomain.banned if the current user is banned or not
+ * chatDomain.channelUnreadCount livedata object returns the number of unread channels for this user
+ * chatDomain.errorEvents events for errors that happen while interacting with the chat
  *
  */
 internal class ChatDomainImpl private constructor(