+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/viewmodel/channels/ChannelsViewModelBinding.kt
@@ -6,7 +6,10 @@ import androidx.lifecycle.LifecycleOwner
 import com.getstream.sdk.chat.view.channels.ChannelsView
 
 @JvmName("bind")
-public fun ChannelsViewModel.bindView(view: ChannelsView, lifecycle: LifecycleOwner) {
+public fun ChannelsViewModel.bindView(
+    view: ChannelsView,
+    lifecycle: LifecycleOwner
+): ChannelsViewModel = apply {
     state.observe(lifecycle) {
         when (it) {
             is ChannelsViewModel.State.Result -> {